New error code and refactor
Somewhere between NetApp 7.x and 9.4 creating a volume that already
existed changed from returning code 17 to 17159. So added that in to the
list and refactored to use the constant names from the netapp library
instead of numbers.

Incidentally, 13130 does not have a documented name but is used when you
try to create an export policy that already exists.

I do not understand the full list of error codes. For example we're
checking for both "volume exists" and "volume does not exist" as
acceptable error codes. Perhaps that's for the delete case (e.g. trying
to delete a volume that doesn't exist is not an error?)

Signed-off-by: Sean Walberg <sean@ertw.com>